djangobackend/pyproject.toml
[project]
name = "djangobackend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "django-extensions>=4.1",
    "djangorestframework>=3.16.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
]
